updated to support APIv4 only
- revise test cases
- merge mattermost_v4 to mattermost, remove all APIv3 endpoints
- the existing webhook-related part was revised to APIv4 endpoints, but not fully tested yet.
