+++ tests/behavior_tests/run_bot.py
@@ -1,13 +1,13 @@
 #!/usr/bin/env python
 from mattermost_bot.bot import Bot, PluginsManager
-from mattermost_bot.mattermost_v4 import MattermostClientv4
+from mattermost_bot.mattermost import MattermostClient
 from mattermost_bot.dispatcher import MessageDispatcher
 import bot_settings
 
 class LocalBot(Bot):
 
     def __init__(self):
-        self._client = MattermostClientv4(
+        self._client = MattermostClient(
             bot_settings.BOT_URL, bot_settings.BOT_TEAM,
             bot_settings.BOT_LOGIN, bot_settings.BOT_PASSWORD,
             bot_settings.SSL_VERIFY
